Move verify_dataset_shares api to dataset_sharing
